7주차 화요일 배포

+++ src/main/java/shop/nuribooks/books/payment/payment/dto/PaymentInfoDto.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public record PaymentInfoDto(
 	BigDecimal totalPrice,
 	BigDecimal unitPrice,
-	BigDecimal shippingPrice,
+	Integer shippingPrice,
 	BigDecimal orderCouponPrice,
 	BigDecimal bookCouponPrice, // 도서 쿠폰 총 금액
 	BigDecimal wrappingPrice,
